requirements.txt
# Web Framework & ASGI Server
fastapi
uvicorn
gunicorn

# Google Generative AI
google-genai

# Database
supabase

# Data Handling & Utilities
pandas
python-dotenv
fuzzywuzzy
py-expression-eval

# Optional: for faster fuzzy string matching
python-levenshtein